Safari-611.1.10/LayoutTests/fast/dom/DOMURL/check-instanceof-domurl-functions.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Test instanceof functions and properties of URL and webkitURL.");

var url = new URL("about:blank");

shouldBeTrue("'createObjectURL' in URL");
shouldBeTrue("'revokeObjectURL' in URL");

shouldBeUndefined("url.createObjectURL");
shouldBeUndefined("url.revokeObjectURL");

shouldBeDefined("window.URL.createObjectURL");
shouldBeDefined("window.URL.revokeObjectURL");

// Following tests remain valid until we actually deprecate webkitURL.
shouldBeTrue("'createObjectURL' in webkitURL");
shouldBeTrue("'revokeObjectURL' in webkitURL");
shouldBeDefined("window.webkitURL.createObjectURL");
shouldBeDefined("window.webkitURL.revokeObjectURL");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
